refactor(interfaces): split interfaces
