+++ apps/webaudio-spectrum-analyser/src/app/classes/analyser/analyser.spec.ts
@@ -1,0 +1,7 @@
+import { Analyser } from './analyser';
+
+describe('Analyser', () => {
+  it('should create an instance', () => {
+    expect(new Analyser()).toBeTruthy();
+  });
+});
