+++ apps/webaudio-spectrum-analyser/src/app/classes/analyser/analyser.ts
@@ -1,0 +1,1 @@
+export class Analyser {}
