+++ apps/webaudio-spectrum-analyser/src/app/classes/pitch-shifter-node/interfaces.ts
@@ -1,0 +1,4 @@
+export interface IPitchShifterNode extends GainNode {
+  shift: number;
+  bufferTime: number;
+}
