+++ apps/webaudio-spectrum-analyser/src/app/classes/worklet-processor/filter-processor.ts
@@ -1,4 +1,4 @@
-import { WorkletFilterParam } from '../../interfaces';
+import { WorkletFilterParam } from './interfaces';
 import { AudioWorkletProcessor } from './util';
 
 export class FilterProcessor extends AudioWorkletProcessor {
