+++ apps/webaudio-spectrum-analyser/src/app/classes/worklet-processor/index.ts
@@ -1,3 +1,4 @@
 export * from './filter-processor';
 export * from './generator-processor';
+export * from './interfaces';
 export * from './util';
