+++ apps/webaudio-spectrum-analyser/src/app/state/audio-graph-ui/interfaces.ts
@@ -1,0 +1,22 @@
+export enum ChartLayout {
+  VERTICAL,
+  HORIZONTAL,
+  GRID,
+  _COUNT,
+}
+
+export enum ChartType {
+  TIME_DOMAIN,
+  FREQUENCY,
+}
+
+export interface SetChartType {
+  index: number;
+  type: ChartType;
+}
+
+export interface AudioGraphUiStateModel {
+  chartCount: number;
+  chartType: ChartType[];
+  chartLayout: ChartLayout;
+}
